refactor: 로그인 에러핸들링 및 컨트롤러 추가
