+++ constants/error.js
@@ -6,7 +6,7 @@ const ERROR = {
   NO_CONTENT: "Content is required.",
   INVALID_EMAIL_FORMAT: "* 이메일 형식에 맞지 않습니다.",
   INVALID_PASSWORD_LENGTH: "* 비밀번호는 최소 8자 이상이어야 합니다.",
-  INVALID_PASSWORD_FORMAT: "* 비밀번호는 최소 한 개 이상의 숫자와 영문자를 포함해야 합니다.",
+  INVALID_PASSWORD_FORMAT: "* 비밀번호는 한 개 이상의 숫자와 영문자를 포함해야 합니다.",
   INVALID_ACCOUNT: "An account with this email does not exist.",
   PASSWORDS_NOT_MATCH: "* 비밀번호가 일치하지 않습니다.",
   WRONG_PASSWORD: "Email and password do not match.",
@@ -14,7 +14,7 @@ const ERROR = {
   GROUP_NAME_DUPLICATE: "This group name is alreday taken.",
   GROUP_APPLICATION_DUPLICATE: "This group has already been applied.",
   CARD_NOT_FOUND: "Card not found.",
-  USER_NOT_FOUND: "User not found.",
+  USER_NOT_FOUND: "* 가입되지 않은 이메일입니다.",
   GROUP_NOT_FOUND: "Group not found.",
   MEMBER_NOT_FOUND: "Member not found.",
   AUTH_FORBIDDEN: "Authentication error.",
