+++ middlewares/validateLogin.js
@@ -1,37 +1,5 @@
-const passport = require("passport");
 const { body } = require("express-validator");
 const ERROR = require("../constants/error");
-
-exports.validateAuth = (req, res, next) => {
-  // passport.authenticate("local", (err, user, info) => {
-  //   if (err) {
-  //     console.log(err);
-  //     const error = new Error("Internal Server Error");
-  //     error.status = 500;
-  //     return next(error);
-  //   }
-  //   if (!user) {
-  //     return res.json("login", { errorMsg: info.message });
-  //   }
-  //   res.locals.user = user;
-  //   next();
-  // })(req, res, next);
-  passport.authenticate("local", (err, user, info) => {
-    if (!user) {
-      return res.json("login", { message: ERROR.INVALID_ACCOUNT });
-    }
-    console.log(111);
-    req.login(user, (err) => {
-      console.log(222);
-
-      if (err) {
-        return next(err);
-      }
-
-      return next();
-    });
-  })(req, res, next);
-};
 
 exports.validateLogin = [
   body("email")
@@ -47,11 +15,4 @@ exports.validateLogin = [
     .withMessage(ERROR.INVALID_PASSWORD_LENGTH)
     .matches(/(?=.*\d)(?=.*[a-zA_Z])/)
     .withMessage(ERROR.INVALID_PASSWORD_FORMAT)
-    .custom((value, { req }) => {
-      if (value !== req.body.passwordConfirm) {
-        return false;
-      }
-      return true;
-    })
-    .withMessage(ERROR.PASSWORDS_NOT_MATCH),
 ];
