+++ routes/login.js
@@ -1,27 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const jwt = require("jsonwebtoken");
-const passport = require("passport");
-const createError = require("http-errors");
 
-const ERROR = require("../constants/error");
+const { validateLogin } = require("../middlewares/validateLogin");
+const { login } = require("./controllers/loginController");
 
-router.post("/", async (req, res, next) => {
-  try {
-    passport.authenticate("local", (err, user) => {
-      if (err || !user) {
-        return res.send(createError(400, ERROR.USER_NOT_FOUND));
-      }
-
-      if (user) {
-        const token = jwt.sign(user.toJSON(), process.env.JWT_SECRET);
-
-        res.json({ user, token: "Bearer " + token });
-      }
-    })(req, res, next);
-  } catch (err) {
-    res.send(createError(400, ERROR.INVALID_ACCOUNT));
-  }
-});
+router.route("/").post(validateLogin, login);
 
 module.exports = router;
